Resolve "Porting on aarch64"

Closes #23

See merge request smartobject/so3!8

+++ so3/include/mmc.h
@@ -251,7 +251,7 @@ struct mmc_data {
 	};
 	uint flags;
 	uint blocks;
-	uint blocksize;
+	unsigned long blocksize;
 };
 
 /* forward decl. */
